save change api movie

+++ src/modules/user/user.controller.ts
@@ -4,38 +4,13 @@ import { Controller, Get, Post, Body, Patch, Param, Delete, Query, UseFilters, U
 import { UserService } from './user.service';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
-import { GenderEnum, UserType } from '@prisma/client';
+import { GenderEnum, RoleEnum } from '@prisma/client';
 import { HttpExceptionFilter, } from 'src/model/http-exception.filter';
 import { BadRequestException, ClassSerializerInterceptor, MiddlewareConsumer, NestModule, NotFoundException, UnauthorizedException } from '@nestjs/common';
 import { SerializeUser, UserEntity } from './entities/user.entity';
 import { Request, Response, NextFunction } from 'express';
 // import { IsAuthMiddleware } from 'src/middlewares/middlewares.service';
 
-
-@Controller('/register')
-export class RegisterController {
-  constructor(
-    private readonly userService: UserService,
-  ) { }
-
-
-  @UseFilters(HttpExceptionFilter)
-  @UseInterceptors(ClassSerializerInterceptor)
-  @Post()
-  async register(@Body() createUserDto: CreateUserDto) {
-    if (!createUserDto.name) throw new BadRequestException('Name is required');
-    if (!createUserDto.email) throw new BadRequestException('Email is required');
-    if (!createUserDto.password) throw new BadRequestException('Password is required');
-    if (createUserDto.password.length < 6) throw new BadRequestException('Password must be at least 6 characters');
-    const isExist = await this.userService.findOneByEmail(createUserDto.email);
-    if (isExist) {
-      throw new BadRequestException('Email already exists');
-    }
-    createUserDto.role = 'USER'
-    const response = await this.userService.createUser(createUserDto);
-    return new UserEntity(response)
-  }
-}
 
 @Controller('user')
 export class UserController {
@@ -53,7 +28,7 @@ export class UserController {
     if(!createUserDto.gender) throw new BadRequestException('Gender is required')
     if(!Object.values(GenderEnum).includes(createUserDto.gender.toUpperCase() as GenderEnum)) throw new BadRequestException('Gender is not valid')
     if (!createUserDto.role) throw new BadRequestException('Role is required');
-    if (!Object.values(UserType).includes(createUserDto.role.toUpperCase() as UserType)) {
+    if (!Object.values(RoleEnum).includes(createUserDto.role.toUpperCase() as RoleEnum)) {
       throw new BadRequestException('Role is not valid');
     }
     if (!createUserDto.password) throw new BadRequestException('Password is required');
@@ -70,20 +45,18 @@ export class UserController {
   @UseFilters(HttpExceptionFilter)
   @UseInterceptors(ClassSerializerInterceptor)
   @Get()
-  async findAllUser(@Req() req: Request, @Query('role') role?: UserType, @Query('gender') gender?: GenderEnum) {
+  async findAllUser(@Req() req: Request, @Query('role') role?: RoleEnum, @Query('gender') gender?: GenderEnum) {
     const userRole = req.payload?.role; // Access the role property from the payload object
     if (userRole !== 'ADMIN') {
       throw new UnauthorizedException('🚫 User is Un-Authorized 🚫')
     }
-    if (role && !Object.values(UserType).includes(role.toUpperCase() as UserType)) {
+    if (role && !Object.values(RoleEnum).includes(role.toUpperCase() as RoleEnum)) {
       throw new BadRequestException()
     }
     if (gender && !Object.values(GenderEnum).includes(gender.toUpperCase() as GenderEnum)) {
       throw new BadRequestException()
     }
     const response = await this.userService.findAllUser(role, gender)
-    console.log(response);
-
     return response.map(user => new UserEntity(user))
   }
 
@@ -104,10 +77,13 @@ export class UserController {
   @UseFilters(HttpExceptionFilter)
   @UseInterceptors(ClassSerializerInterceptor)
   @Delete(':id')
-  async deleteOne(@Param('id') id: number) {
+  async deleteOne(@Req() req: Request, @Param('id') id: number) {
+    const userRole = req.payload.role
+    if(userRole !== 'ADMIN'){
+      throw new UnauthorizedException('🚫 User is Un-Authorized 🚫')
+    }
     const existingUser = await this.userService.findUserById(+id)
     if (!existingUser) throw new BadRequestException()
-
     const user = await this.userService.deleteUser(+id)
     return new UserEntity(user)
   }
